mesh-agents/BlockBeak-TelegramBot/pyproject.toml
[project]
name = "BlockBeak-TelegramBot"
version = "0.1.0"
description = "BlockBeak Telegram Bot"
authors = [
    { name = "Heurist Team", email = "team@heurist.xyz" },
]
dependencies = [
    "python-dotenv==1.0.0",
    "pyTelegramBotAPI>=4.14.0",
    "openai-agents[litellm]==0.0.14",
    "pyyaml>=6.0.1",
]
requires-python = ">=3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"] 
